requirements.txt
streamlit==1.28.1
Jinja2==3.1.2
Pillow==10.1.0
python-dateutil==2.8.2
requests==2.31.0
PyMuPDF==1.23.8
